Refactor validation criteria and validation items
