+++ nomenclature/processor/__init__.py
@@ -5,5 +5,5 @@ from nomenclature.processor.region import (  # noqa
 )
 from nomenclature.processor.nuts import NutsProcessor  # noqa
 from nomenclature.processor.required_data import RequiredDataValidator  # noqa
-from nomenclature.processor.data_validator import DataValidator  # noqa
+from nomenclature.processor.data import DataValidator  # noqa
 from nomenclature.processor.aggregator import Aggregator  # noqa
