+++ nomenclature/processor/data.py
@@ -7,18 +7,11 @@ import pandas as pd
 import yaml
 from pyam import IamDataFrame
 from pyam.utils import adjust_log_level
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Field,
-    computed_field,
-    field_validator,
-    model_validator,
-)
 
 from nomenclature.codelist import VariableCodeList
 from nomenclature.definition import DataStructureDefinition
 from nomenclature.exceptions import DataValidationError, NoTracebackExceptionGroup
+from nomenclature.processor.validator import ValidationItem
 from nomenclature.processor import Processor
 from nomenclature.processor.iamc import IamcDataFilter
 from nomenclature.utils import get_relative_path
@@ -33,169 +26,18 @@ class WarningEnum(IntEnum):
     low = 20
 
 
-class DataValidationCriteria(BaseModel):
-    warning_level: WarningEnum = WarningEnum.error
-
-    model_config = ConfigDict(extra="forbid")
-
-    @field_validator("warning_level", mode="before")
-    @classmethod
-    def validate_warning_level(cls, value):
-        if isinstance(value, str):
-            try:
-                return WarningEnum[value]
-            except KeyError:
-                raise ValueError(
-                    f"Invalid warning level: {value}. Expected one of:"
-                    f" {', '.join(level.name for level in WarningEnum)}"
-                )
-        return value
-
-    @property
-    def criteria(self):
-        pass
-
-    def __str__(self):
-        return ", ".join([f"{key}: {value}" for key, value in self.criteria.items()])
-
-
-class DataValidationValue(DataValidationCriteria):
-    value: float
-    rtol: float = 0.0
-    atol: float = 0.0
-
-    @property
-    def tolerance(self) -> float:
-        return self.value * self.rtol + self.atol
-
-    @computed_field
-    def upper_bound(self) -> float:
-        return self.value + self.tolerance
-
-    @computed_field
-    def lower_bound(self) -> float:
-        return self.value - self.tolerance
-
-    @property
-    def validation_args(self):
-        """Attributes used for validation (as bounds)."""
-        return self.model_dump(
-            exclude_none=True,
-            exclude_unset=True,
-            exclude=["warning_level", "value", "rtol", "atol"],
-        )
-
-    @property
-    def criteria(self):
-        """Attributes used for validation (as specified in the file)."""
-        return self.model_dump(
-            exclude_none=True,
-            exclude_unset=True,
-            exclude=["warning_level", "lower_bound", "upper_bound"],
-        )
-
-
-class DataValidationBounds(DataValidationCriteria):
-    upper_bound: float | None = None
-    lower_bound: float | None = None
-
-    # Allow extra but raise error to guard against multiple criteria
-    model_config = ConfigDict(extra="allow")
-
-    @model_validator(mode="after")
-    def check_validation_criteria_exist(self):
-        if self.upper_bound is None and self.lower_bound is None:
-            raise ValueError("No validation criteria provided: " + str(self.criteria))
-        return self
-
-    @model_validator(mode="after")
-    def check_validation_multiple_criteria(self):
-        if self.model_extra:
-            raise ValueError(
-                "Must use either bounds, range or value, found: " + str(self.criteria)
-            )
-        return self
-
-    @property
-    def validation_args(self):
-        return self.criteria
-
-    @property
-    def criteria(self):
-        return self.model_dump(
-            exclude_none=True, exclude_unset=True, exclude=["warning_level"]
-        )
-
-
-class DataValidationRange(DataValidationCriteria):
-    range: list[float] = Field(..., min_length=2, max_length=2)
-
-    @field_validator("range", mode="after")
-    @classmethod
-    def check_range_is_valid(cls, value: list[float]):
-        if value[0] > value[1]:
-            raise ValueError(
-                "Validation 'range' must be given as `(lower_bound, upper_bound)`, "
-                "found: " + str(value)
-            )
-        return value
-
-    @computed_field
-    def upper_bound(self) -> float:
-        return self.range[1]
-
-    @computed_field
-    def lower_bound(self) -> float:
-        return self.range[0]
-
-    @property
-    def validation_args(self):
-        """Attributes used for validation (as bounds)."""
-        return self.model_dump(
-            exclude_none=True,
-            exclude_unset=True,
-            exclude=["warning_level", "range"],
-        )
-
-    @property
-    def criteria(self):
-        return self.model_dump(
-            exclude_none=True,
-            exclude_unset=True,
-            exclude=["warning_level", "lower_bound", "upper_bound"],
-        )
-
-
-class DataValidationItem(IamcDataFilter):
-    name: str | None = None
-    validation: list[DataValidationValue | DataValidationRange | DataValidationBounds]
-
-    @model_validator(mode="after")
-    def check_warnings_order(self):
-        """Check if warnings are set in descending order of severity."""
-        if self.validation != sorted(
-            self.validation, key=lambda c: c.warning_level, reverse=True
-        ):
-            raise ValueError(
-                f"Validation criteria for {self.criteria} not sorted"
-                " in descending order of severity."
-            )
-        else:
-            return self
-
+class DataValidationItem(ValidationItem, IamcDataFilter):
     @property
     def filter_args(self):
-        """Attributes used for validation (as specified in the file)."""
+        """Dimensions and values used to filter rows to be validated."""
         return self.model_dump(
             exclude_none=True, exclude_unset=True, exclude=["validation", "name"]
         )
 
-    def __str__(self):
-        return ", ".join([f"{key}: {value}" for key, value in self.filter_args.items()])
-
     def apply(
         self, df: IamDataFrame, fail_list: list, output_list: list
     ) -> tuple[bool, list, list]:
+        """Apply validation to IamDataFrame."""
         error = False
         per_item_df = df.filter(**self.filter_args)
